DW's patch plus a small fix to allow code to run on Mahti.

+++ main.c
@@ -65,7 +65,7 @@ int main(int argc, char* argv[])
     }
 
     // How big is the system
-    p.N = p.Lx * p.Ly * p.Lz;
+    p.N = ((long long) p.Lx) * ((long long) p.Ly) * ((long long) p.Lz);
 
     // Scale factor (fixed to 1 for now, no expansion)
     p.a = 1.0;
